5. TODO list [#3786 #321234] + Account update

+++ src/main/java/ru/job4j/todo/controller/ItemController.java
@@ -3,9 +3,13 @@ package ru.job4j.todo.controller;
 import net.jcip.annotations.ThreadSafe;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ModelAttribute;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
 import ru.job4j.todo.model.Account;
 import ru.job4j.todo.model.Item;
+import ru.job4j.todo.service.AccountService;
 import ru.job4j.todo.service.ItemServiceImpl;
 import ru.job4j.todo.utility.Utility;
 import javax.servlet.http.HttpSession;
@@ -15,9 +19,11 @@ import java.util.Date;
 @ThreadSafe
 public class ItemController {
     ItemServiceImpl itemService;
+    AccountService accountService;
 
-    public ItemController(ItemServiceImpl itemService) {
+    public ItemController(ItemServiceImpl itemService, AccountService accountService) {
         this.itemService = itemService;
+        this.accountService = accountService;
     }
 
     @GetMapping("/index")
@@ -47,14 +53,16 @@ public class ItemController {
     @GetMapping("/formAddItem")
     public String addItem(Model model, HttpSession session) {
         model.addAttribute("", new Item("name", "description",
-                new Date(), true));
+                new Date(), true, new Account()));
         Account account = Utility.logAccount(session);
         model.addAttribute("account", account);
         return "createItem";
     }
 
     @PostMapping("/createItem")
-    public String createItem(@ModelAttribute Item item) {
+    public String createItem(@ModelAttribute Item item, HttpSession session) {
+        Account account = Utility.logAccount(session);
+        item.setAccount(account);
         itemService.add(item);
         return "redirect:/index";
     }
@@ -66,9 +74,13 @@ public class ItemController {
     }
 
     @PostMapping("/updateItem")
-    public String updateItem(@ModelAttribute Item item) {
-        itemService.replace(item.getId(), item);
-        return "redirect:/index";
+    public String updateItem(@ModelAttribute Item item, HttpSession session) {
+        Account account = Utility.logAccount(session);
+        if (account.getName().equals(itemService.findById(item.getId()).getAccount().getName())) {
+            itemService.replace(item.getId(), item);
+            return "redirect:/index";
+        }
+        return "failupdate";
     }
 
     @GetMapping("/itemProfile/{itemId}")
@@ -80,14 +92,22 @@ public class ItemController {
     }
 
     @GetMapping("/execute/{itemId}")
-    public String executedItem(@PathVariable("itemId") int id) {
-        itemService.executeById(id);
-        return "redirect:/index";
+    public String executedItem(@PathVariable("itemId") int id, HttpSession session) {
+        Account account = Utility.logAccount(session);
+        if (account.getName().equals(itemService.findById(id).getAccount().getName())) {
+            itemService.executeById(id);
+            return "redirect:/index";
+        }
+        return "failexecute";
     }
 
     @GetMapping("/deleteItem/{itemId}")
-    public String deleteItem(@PathVariable("itemId") int id) {
-        itemService.delete(id);
-        return "redirect:/index";
+    public String deleteItem(@PathVariable("itemId") int id, HttpSession session) {
+        Account account = Utility.logAccount(session);
+        if (account.getName().equals(itemService.findById(id).getAccount().getName())) {
+            itemService.delete(id);
+            return "redirect:/index";
+        }
+        return "faildelete";
     }
 }
